+++ packages/storybook/src/nimble/mapping/text/mapping-text.stories.ts
@@ -1,7 +1,10 @@
 import { html } from '@microsoft/fast-element';
 import type { Meta, StoryObj } from '@storybook/html';
 import { mappingKeyDescription } from '../base/story-helpers';
-import { apiCategory, createUserSelectedThemeStory } from '../../../utilities/storybook';
+import {
+    apiCategory,
+    createUserSelectedThemeStory
+} from '../../../utilities/storybook';
 import { hiddenWrapper } from '../../../utilities/hidden';
 
 const metadata: Meta = {
@@ -24,11 +27,11 @@ export const textMapping: StoryObj = {
         key: {
             description: mappingKeyDescription('the mapped display text'),
             control: false,
-            table: { category: apiCategory.attributes },
+            table: { category: apiCategory.attributes }
         },
         text: {
             description: 'The display text.',
-            table: { category: apiCategory.attributes },
+            table: { category: apiCategory.attributes }
         }
     },
     args: {}